+++ regserver/regulations/tests/views_chrome_tests.py
@@ -26,3 +26,24 @@ class ViewsChromeTest(TestCase):
         view.request = RequestFactory().get('/')
         response = view.get(view.request, label_id='lab', version='ver')
         self.assertEqual(404, response.status_code)
+
+    @patch('regulations.views.chrome.ChromeView.set_tree_context')
+    @patch('regulations.views.chrome.generator')
+    def test_error_propagation(self, generator, set_tree_context):
+        """Test that the response of the outer view is that of the inner
+        when there's an error"""
+        generator.get_regulation.return_value = {}
+        generator.get_tree_paragraph.return_value = {}
+        set_tree_context.return_value = None
+
+        class InnerView(TemplateView):
+            def get(self, request, *args, **kwargs):
+                return HttpResponseGone()
+
+        class FakeView(ChromeView):
+            partial_class = InnerView
+
+        view = FakeView()
+        view.request = RequestFactory().get('/')
+        response = view.get(view.request, label_id='lab', version='ver')
+        self.assertEqual(410, response.status_code)
