+++ regserver/regulations/tests/views_with_chrome_tests.py
@@ -13,27 +13,6 @@ class ViewTests(TestCase):
     def tearDown(self):
         settings.DEBUG = self.original_debug
 
-    def test_add_extras(self):
-        view = ChromeView()
-        context = {}
-        view.add_extras(context)
-
-        self.assertTrue('GOOGLE_ANALYTICS_ID' in context)
-        self.assertTrue('GOOGLE_ANALYTICS_SITE' in context)
-        self.assertTrue('env' in context)
-
-    def test_add_extras_env(self):
-        view = ChromeView()
-        context = {}
-        
-        settings.DEBUG = True
-        view.add_extras(context)
-        self.assertEqual('source', context['env'])
-
-        settings.DEBUG = False
-        view.add_extras(context)
-        self.assertEqual('built', context['env'])
-
     @patch('regulations.views.chrome.generator')
     def test_get_404(self, generator):
         generator.get_regulation.return_value = None
